Added message cache
TODO: Clean up the comments

+++ src/main/java/me/nopox/utils/storage/repository/redis/RedisRepository.java
@@ -72,7 +72,7 @@ public abstract class RedisRepository<K extends String, T> implements Repository
         });
     }
 
-    public void saveExpireable(K key, T value, int seconds) {
+    public void saveExpireable(K key, T value, long seconds) {
         CompletableFuture.runAsync(() -> {
             Jedis jedis = this.jedis.getJedisResource();
